feat: drop support for Angular workspaces
We have decided to stop supporting Angular Workspaces and focus exclusively on Nx Workspaces. This
decision is motivated by several factors, including the fact that the Nx Devkit no longer supports
compatibility with Angular Workspaces. To offer support, we would have to create a second plugin
that uses AngularDevkit under the hood, which would require considerable effort on the part of our
maintainers. We want to focus on what brings the most value to our current users, and Nx Workspaces
is the best way to do that.

BREAKING CHANGE: As an alternative, maintainers suggest migrating your Angular Workspaces to Nx
Workspaces if you want to use this plugin. A stand-alone approach to Nx workspaces is an option that
you should consider. We apologize for any inconvenience this may cause and appreciate your
understanding.

close #467